+++ src/sampleApp/phusnyd/GenerateBox/BoxGenerateApp.jsx
@@ -3,12 +3,14 @@ import GenerateBox from "./components/GenerateBox";
 import { useState } from "react";
 
 function BoxGenerateApp() {
-  const [boxes, setBoxes] = useState("");
-  const [showBoxes, setShowBoxes] = useState();
+  const [boxes, setBoxes] = useState(0);
+  const [showBoxes, setShowBoxes] = useState([]);
+
   function handleSubmit(e) {
     e.preventDefault();
-    setShowBoxes((prevState) => (prevState = boxes));
-    setBoxes("");
+    const boxsArray = Array.from({ length: boxes }, (_, i) => i + 1);
+    setShowBoxes(boxsArray);
+    setBoxes(0);
   }
 
   return (
