+++ src/sampleApp/phusnyd/GenerateBox/components/BoxList.jsx
@@ -1,34 +1,19 @@
 import { useState } from "react";
+import BoxItem from "./BoxItem";
 
 function BoxList({ showBoxes }) {
-  var numberArr = Array.from({ length: showBoxes }, (_, i) => i + 1);
-  const [changeColor, setChangeColor] = useState("#fffff");
-  function handleChangeColor() {
-    const randomColor = "#" + Math.random().toString(16).slice(2, 8);
-    setChangeColor((preState) => (preState = randomColor));
-    console.log("color: ", changeColor);
-  }
-
   return (
     <>
-      {showBoxes ? (
+      {showBoxes.length > 0  ? (
         <ul className="boxList">
-          {numberArr.length > 0 ? (
-            numberArr.map((number, index) => {
-              return (
-                <li
-                  key={index}
-                  style={{ backgroundColor: `${changeColor}` }}
-                  className="box"
-                  onClick={handleChangeColor}
-                >
-                  box #{number}
-                </li>
-              );
-            })
-          ) : (
-            <p>Nothing in Array</p>
-          )}
+         {showBoxes.map((number, index) => {
+            return (
+              <BoxItem 
+                key={index}
+                number={number}
+              />
+            );
+          })}
         </ul>
       ) : (
         <p>No Box!</p>
